Improve test mockery and avoid deprecated RLContext::getConfig
ResourceLoaderContext::getConfig is being deprecated in MediaWiki
(I74a9535918ea43b). The extension should use a Config instance
injected into the Module class instead.

Improve other mocking while at it, such as injecting the 'skin'
parameter with a mock (instead of the indirection through global
stuff and incidental validation), avoid database interaction
to create an '!User::isAnon()' object.

Keep one case as full integration test, moved to a separate class.
This was incompatible with the new structure due to it relying
on changing global settings (wgLanguageCode), which is ignored
by the code directly being tested so it would pass for the wrong
reasons if kept as-is.

Change-Id: I2cc52f5308081dda469a8a654ddc273a56613677