Fixes mismatch for default configuration
When the extension is installed with the default configuration the
ResourceLoader module is not registered but the module is added
to the page. This leads to problems in integration tests which
are expecting modules added to the page to exist.

This change unifies the logic between these two hooks so that
we don't end up in that scenario.

Change-Id: I0c8a0f4b23bce248fa8c2385cf59b7db4233ee37