Do not show help links for temp accounts
Although this extension already disallows editing user CSS and JS pages
for temporary accounts, the help link on the user pages still shows up
if you try to visit them with a temporary account, while an anonymous
user does not get this help link. This patch replaces a couple of
isRegistered() checks to isNamed() to fix this.

Bug: T326919
Change-Id: I9cacfc5de57be73663f8b808d6a05e531ffdbad7